Atualizando classe de entidade dono, onde o campo CPF/CNPJ aceita valores nulos

+++ humv_backend/src/main/java/br/edu/ufrb/lasis/humv/dao/SetorDAO.java
@@ -80,7 +80,7 @@ public class SetorDAO  extends GenericDAO<Setor> implements Serializable {
 	public List<Setor> search(String palavrachave) {
 		Criteria criteria = getCriteria();
 
-		Integer conversionResult = NumberUtils.convertStringToInteger(palavrachave);
+		BigInteger conversionResult = NumberUtils.convertStringToInteger(palavrachave);
 		if (conversionResult != null) {
 			criteria.add(
 					Restrictions.or(
